scraper/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.54
easyocr>=1.7.0
Pillow>=10.0.0
keyboard>=0.13.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
